Adjusted test names for meta
